+++ app/controllers/toilets.js
@@ -1,7 +1,14 @@
+const { check } = require('express-validator/check');
+const { getValidationState } = require('../utils/validationHelper');
 const Toilets = require('../models/toilets');
 const { convertPointToGeoJSONPoint } = require('../utils/geo');
 
-module.exports.getToilets = (req, res) => {
+exports.getToilets = function getToilets(req, res) {
+  const state = getValidationState(req);
+  if (state.hasErrors) {
+    return res.status(401).json({ errors: state.errors });
+  }
+
   Toilets.find()
     .limit(500)
     .exec((err, toilets) => {
@@ -12,7 +19,12 @@ module.exports.getToilets = (req, res) => {
     });
 };
 
-module.exports.getToiletById = (req, res) => {
+exports.getToiletById = function getToiletById(req, res) {
+  const state = getValidationState(req);
+  if (state.hasErrors) {
+    return res.status(401).json({ errors: state.errors });
+  }
+
   const { id } = req.params;
 
   Toilets.findById(id, (err, toilet) => {
@@ -24,7 +36,12 @@ module.exports.getToiletById = (req, res) => {
   });
 };
 
-module.exports.createToilet = (req, res) => {
+exports.createToilet = function createToilet(req, res) {
+  const state = getValidationState(req);
+  if (state.hasErrors) {
+    return res.status(401).json({ errors: state.errors });
+  }
+
   const {
     title,
     location,
@@ -48,7 +65,12 @@ module.exports.createToilet = (req, res) => {
   });
 };
 
-module.exports.updateToilet = (req, res) => {
+exports.updateToilet = function updateToilet(req, res) {
+  const state = getValidationState(req);
+  if (state.hasErrors) {
+    return res.status(401).json({ errors: state.errors });
+  }
+
   const {
     id,
     title,
@@ -75,7 +97,12 @@ module.exports.updateToilet = (req, res) => {
   });
 };
 
-module.exports.deleteToilet = (req, res) => {
+exports.deleteToilet = function deleteToilet(req, res) {
+  const state = getValidationState(req);
+  if (state.hasErrors) {
+    return res.status(401).json({ errors: state.errors });
+  }
+
   const { id } = req.params;
 
   Toilets.findByIdAndDelete(id, (err) => {
@@ -87,8 +114,13 @@ module.exports.deleteToilet = (req, res) => {
   });
 };
 
-module.exports.deleteManyToilets = (req, res) => {
-  const ids = req.body;
+exports.deleteManyToilets = function deleteManyToilets(req, res) {
+  const state = getValidationState(req);
+  if (state.hasErrors) {
+    return res.status(401).json({ errors: state.errors });
+  }
+
+  const { ids } = req.body;
 
   Toilets.deleteMany({
     id: { $in: ids },
@@ -99,4 +131,54 @@ module.exports.deleteManyToilets = (req, res) => {
 
     return res.status(204);
   });
+};
+
+exports.validate = (method) => {
+  switch (method) {
+    case exports.getToilets.name: {
+      return [];
+    }
+    case exports.getToiletById.name: {
+      return [
+        check('id').exists().isMongoId(),
+      ];
+    }
+    case exports.createToilet.name: {
+      return [
+        check('title').optional(),
+        check('location').exists().isArray().isLength({ min: 2, max: 2 }),
+        check('author').exists().isString().not()
+          .isEmpty(),
+        check('address').exists().isString().not()
+          .isEmpty(),
+        check('description').optional().isString(),
+      ];
+    }
+    case exports.updateToilet.name: {
+      return [
+        check('id').exists().isMongoId(),
+        check('title').optional(),
+        check('location').exists().isArray().isLength({ min: 2, max: 2 }),
+        check('author').optional().isString().not()
+          .isEmpty(),
+        check('address').exists().isString().not()
+          .isEmpty(),
+        check('description').optional().isString(),
+      ];
+    }
+    case exports.deleteToilet.name: {
+      return [
+        check('id').exists().isMongoId(),
+      ];
+    }
+    case exports.deleteManyToilets.name: {
+      return [
+        check('ids').exists().isArray().isLength({ min: 1 }),
+      ];
+    }
+
+    default: {
+      return [];
+    }
+  }
 };
